Add configuration for bmc dump
By default, the rshim is disabled in the BMC.
The bmc dump in BF can start the rshim and is
also creating a very large dump.
If the dump is triggered autonomously, in case of restart the
server, the rshim may be taken by the BMC instead of the server.

Test:
Reboot BMC and verify that the rshim is not taken.

Fixes nvbug https://redmine.mellanox.com/issues/4092772

Signed-off-by: afogel <afogel@afogel2.nvidia.com>